+++ app/upgrades/v15/upgrades.go
@@ -41,6 +41,7 @@ func CreateUpgradeHandler(
 		poolmanagerParams.PoolCreationFee = keepers.GAMMKeeper.GetParams(ctx).PoolCreationFee
 
 		keepers.PoolManagerKeeper.SetParams(ctx, poolmanagerParams)
+		//nolint:errcheck
 		keepers.PacketForwardKeeper.SetParams(ctx, packetforwardtypes.DefaultParams())
 		setICQParams(ctx, keepers.ICQKeeper)
 
@@ -77,6 +78,7 @@ func setICQParams(ctx sdk.Context, icqKeeper *icqkeeper.Keeper) {
 	icqparams.AllowQueries = wasmbinding.GetStargateWhitelistedPaths()
 	// Adding SmartContractState query to allowlist
 	icqparams.AllowQueries = append(icqparams.AllowQueries, "/cosmwasm.wasm.v1.Query/SmartContractState")
+	//nolint:errcheck
 	icqKeeper.SetParams(ctx, icqparams)
 }