+++ app/upgrades/v21/upgrades.go
@@ -1,9 +1,7 @@
 package v21
 
 import (
-	"cosmossdk.io/math"
 	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
-	auctiontypes "github.com/skip-mev/block-sdk/x/auction/types"
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -12,6 +10,8 @@ import (
 	icacontrollertypes "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/controller/types"
 	icahosttypes "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/host/types"
 	ibctransfertypes "github.com/cosmos/ibc-go/v7/modules/apps/transfer/types"
+
+	icqtypes "github.com/cosmos/ibc-apps/modules/async-icq/v7/types"
 
 	"github.com/osmosis-labs/osmosis/osmoutils"
 	"github.com/osmosis-labs/osmosis/v20/app/keepers"
@@ -94,15 +94,12 @@ func CreateUpgradeHandler(
 				keyTable = icahosttypes.ParamKeyTable() //nolint:staticcheck
 			case icacontrollertypes.SubModuleName:
 				keyTable = icacontrollertypes.ParamKeyTable() //nolint:staticcheck
+			case icqtypes.ModuleName:
+				keyTable = icqtypes.ParamKeyTable() //nolint:staticcheck
 
 			// wasm
 			case wasmtypes.ModuleName:
 				keyTable = wasmtypes.ParamKeyTable() //nolint:staticcheck
-
-			// POB
-			case auctiontypes.ModuleName:
-				// already SDK v47
-				continue
 
 			// osmosis modules
 			case protorevtypes.ModuleName:
@@ -150,15 +147,10 @@ func CreateUpgradeHandler(
 			return nil, err
 		}
 
-		// Set POB Params:
-		err = setAuctionParams(ctx, keepers)
-		if err != nil {
-			return nil, err
-		}
-
 		// Set expedited proposal param:
 		govParams := keepers.GovKeeper.GetParams(ctx)
 		govParams.ExpeditedMinDeposit = sdk.NewCoins(sdk.NewCoin("uosmo", sdk.NewInt(5000000000)))
+		govParams.MinInitialDepositRatio = "0.250000000000000000"
 		err = keepers.GovKeeper.SetParams(ctx, govParams)
 		if err != nil {
 			return nil, err
@@ -186,18 +178,4 @@ func CreateUpgradeHandler(
 
 		return migrations, nil
 	}
-}
-
-func setAuctionParams(ctx sdk.Context, keepers *keepers.AppKeepers) error {
-	pobAddr := keepers.AccountKeeper.GetModuleAddress(auctiontypes.ModuleName)
-
-	auctionParams := auctiontypes.Params{
-		MaxBundleSize:          2,
-		EscrowAccountAddress:   pobAddr,
-		ReserveFee:             sdk.Coin{Denom: "uosmo", Amount: sdk.NewInt(1_000_000)},
-		MinBidIncrement:        sdk.Coin{Denom: "uosmo", Amount: sdk.NewInt(1_000_000)},
-		FrontRunningProtection: true,
-		ProposerFee:            math.LegacyNewDecWithPrec(25, 2),
-	}
-	return keepers.AuctionKeeper.SetParams(ctx, auctionParams)
 }