+++ x/poolmanager/keeper.go
@@ -117,10 +117,30 @@ func (k Keeper) InitGenesis(ctx sdk.Context, genState *types.GenesisState) {
 	} else {
 		k.SetTakerFeeTrackerStartHeight(ctx, ctx.BlockHeight())
 	}
+
+	// Reset the pool volumes KVStore.
+	for _, poolVolume := range genState.PoolVolumes {
+		k.SetVolume(ctx, poolVolume.PoolId, poolVolume.PoolVolume)
+	}
 }
 
 // ExportGenesis returns the poolmanager module's exported genesis.
 func (k Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
+	pools, err := k.AllPools(ctx)
+	if err != nil {
+		panic(err)
+	}
+
+	// Utilize poolVolumes struct to export pool volumes from KVStore.
+	poolVolumes := make([]*types.PoolVolume, len(pools))
+	for i, pool := range pools {
+		poolVolume := k.GetTotalVolumeForPool(ctx, pool.GetId())
+		poolVolumes[i] = &types.PoolVolume{
+			PoolId:     pool.GetId(),
+			PoolVolume: poolVolume,
+		}
+	}
+
 	// Export KVStore values to the genesis state so they can be imported in init genesis.
 	takerFeesTracker := types.TakerFeesTracker{
 		TakerFeesToStakers:         k.GetTakerFeeTrackerForStakers(ctx),
@@ -132,6 +152,7 @@ func (k Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
 		NextPoolId:       k.GetNextPoolId(ctx),
 		PoolRoutes:       k.getAllPoolRoutes(ctx),
 		TakerFeesTracker: &takerFeesTracker,
+		PoolVolumes:      poolVolumes,
 	}
 }